Extract `CommitmentMessageId` into its own module and add a unit test
