+++ src/bridge/client/client.rs
@@ -14,6 +14,7 @@ use std::{
 };
 
 use crate::bridge::{
+    commitments::CommitmentMessageId,
     connectors::{
         base::TaprootConnector,
         connector_0::Connector0,
@@ -29,7 +30,7 @@ use crate::bridge::{
             REWARD_MULTIPLIER, REWARD_PRECISION,
         },
         peg_in::{PegInDepositorStatus, PegInVerifierStatus},
-        peg_out::{CommitmentMessageId, PegOutOperatorStatus},
+        peg_out::PegOutOperatorStatus,
     },
     scripts::generate_pay_to_pubkey_script_address,
     transactions::{
