+++ tests/bridge/integration/peg_out/start_time.rs
@@ -1,7 +1,8 @@
 use bitcoin::{Address, Amount, OutPoint};
 use bitvm::bridge::{
+    commitments::CommitmentMessageId,
     connectors::base::TaprootConnector,
-    graphs::{base::DUST_AMOUNT, peg_out::CommitmentMessageId},
+    graphs::base::DUST_AMOUNT,
     superblock::get_start_time_block_number,
     transactions::{
         base::{BaseTransaction, Input, MIN_RELAY_FEE_KICK_OFF_1, MIN_RELAY_FEE_START_TIME},
